moved query host zone
